Extra Credit: Prime Factorization
WARNING: Using an Eratosthene sieve for prime generation and trial division for factorization. This implementation has time complexity that is exponential in n, and should not be used to factor n > ~2^36.

Finding largest prime factor of n = 2^36-1 = 68719476735 (191) took 33.581s in Chrome on 1.7GHz processor, with majority of time spent generating primes.